LLM/pyproject.toml
[project]
name = "eec"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.3.27",
    "langchain-groq>=0.3.8",
    "python-dotenv>=1.2.1",
]
